+++ CHANGELOG.md
@@ -2,9 +2,14 @@
 
 ## 2.0.0 - 2024-05-01
 
-- Balloon Birds now have parachutes
+- Added players HUD during battle
+- Characters can have backup balloons
+- Characters with backup balloons can refill balloons while on ground
+- Added setup option to start with no balloons
+- Added parachute ability for some characters
 - Characters can walk and jump without balloons
 - Colliding with a character with no balloons kills them
+- Characters can now start with up to 99 balloons
 
 ## 1.0.0 - 2024-04-29